+++ app/src/main/java/ar/edu/ort/parcial/screens/auth/ForgotNewPassword.kt
@@ -1,4 +1,4 @@
-package ar.edu.ort.parcial.screens
+package ar.edu.ort.parcial.screens.auth
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -30,11 +30,13 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 
-@Composable
-fun ForgotPassword() {
-    var email by remember { mutableStateOf("") }
 
-    val isButtonEnabled = email.isNotBlank()
+@Composable
+fun ForgotNewPassword() {
+    var email by remember { mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+
+    val isButtonEnabled = email.isNotBlank() && password.isNotBlank()
 
     Box(
         modifier = Modifier
@@ -71,6 +73,12 @@ fun ForgotPassword() {
                 value = email,
                 onValueChange = { email = it }
             )
+            Spacer(modifier = Modifier.height(18.dp))
+            FieldCom(
+                text = stringResource(id = R.string.field_pass),
+                value = password,
+                onValueChange = { password = it }
+            )
         }
         Column(
             modifier = Modifier
@@ -95,7 +103,6 @@ fun ForgotPassword() {
 
 @Preview()
 @Composable
-fun ForgotPasswordPreview() {
-    ForgotPassword()
+fun ForgotNewPasswordPreview() {
+    ForgotNewPassword()
 }
-
