+++ app/src/main/java/ar/edu/ort/parcial/screens/home/Home.kt
@@ -1,4 +1,4 @@
-package ar.edu.ort.parcial.screens
+package ar.edu.ort.parcial.screens.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
