+++ app/src/main/java/ar/edu/ort/parcial/screens/profile/SellerScreen.kt
@@ -1,8 +1,7 @@
-package ar.edu.ort.parcial.screens
+package ar.edu.ort.parcial.screens.profile
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.shape.CircleShape
