+++ app/src/main/java/ar/edu/ort/parcial/screens/settings/SettingsScreen.kt
@@ -1,4 +1,3 @@
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
